+++ src/main/java/com/dhulai/repository/ServiceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dhulai.entity.ServicesWash;
 
-public interface ServiceRepository extends JpaRepository<ServicesWash, Integer> {
+public interface ServiceRepository extends JpaRepository<ServicesWash, Long> {
     // Add custom queries or methods if needed
 }
